Add `hanami server` command (#22)
The server is just a thin wrapper on top of Rack::Server. The options
that it accepts fall into two different categories:

- When not explicitly set, port and host are not passed to the rack
server instance. This way, they can be configured through the configured
rack handler (e.g., the puma configuration file).

- All others are always given by the Hanami command.

The test for the command forks a new process so that the one running the
tests is not blocked and can assert that endpoint is working. Standard
outputs are silenced to avoid the command output being rendered.