Use Hanami.app_path to find if we're in an app
This allows the hanami CLI to expose its app commands even if invoked from a deeper app directory.

Since this may have caused an issue with the CLI properly executing in earlier testing (before our 2.0.0.beta3 release), wrap the call to `Hanami.app_path` in a rescue that will fall back to a simpler mechanism that will still work in the common cases.