Fix db rollback command
The current implementation of the rollback command does not work. The
target that it passes to the run_migrations method of the database util
class is the current migration which results in a no-op.

This change addresses this problem by explicitly targeting the second
last applied migration. This change also refactors the logic for
when no migrations exist to rollback to provide a clearer message.