Remove unused assets settings (#123)
We're not using the `entry_points` and `destination` assets settings right now. These are currently hardcoded inside assets-js, so remove these references to avoid any confusion that we may be able to configure these from the outside. 

While we're here, properly shellescape the `sri` argument given to the assets CLI.