Allow --skip-db and --database to be given, preferring --skip-db (#266)
Raising a conflicting options error when both are given is a problem, because we _default_ the `database` option to be “sqlite”, which means it’s impossible for the user to un-set.

Instead or raising an error, simply allow `—skip-db` to take priority when given. This means if it is given together with `--database`, the `--database` option will be ignored.