Clarify customizability of config/assets.js
Previously, there was a line that was just `await assets.run()` that was not commented, and a commented block that was suggested for customization. As a result, it was possible that the initial line would be left when enabling the block, which causes a double-build and can also cause a lot of confusion once some customization is added. 

Given that the default customization function doesn't do anything, the computational cost of just executing it is almost certainly acceptable to simplify the setup and reduce the chance of frustration.