Disallow `foreman`'s environment setting. (#305)
Foreman will process the `.env` file by default, but
it does so in a way that is incompatible with the envvar and
shell expansions that `dotenv` provides.

setting `--env=/dev/null` overrides foreman's attempts to
load the environment, allowing `dotenv` to handle the environment
as it would if running `hanami server` directly.

ref: https://github.com/ddollar/foreman/issues/702

fixes: #304