Inject exit command to Create in tests (#307)
There was an issue with test runs ending abruptly with no message when a
database configuration was incorrect. In such cases the Create command
ran an `exit(1)`, which exited not only the command, but the whole test
run.

The solution was already prepared, but not used: it was already possible
to inject an alternative to just calling the actual `Kernel.exit`. This
injects a test double instead and asserts it is not called.